fix: dashboard components is splitted into different files
