+++ src/Routes/Dashboard/Dashboard.css
@@ -1,0 +1,4 @@
+.board{
+    min-height: 60vh;
+    min-width: 80vw;
+}
